python/tensogram-earthkit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tensogram-earthkit"
version = "0.18.1"
description = "earthkit-data source and encoder plugins for tensogram .tgm files"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [{name = "ECMWF", email = "software@ecmwf.int"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
dependencies = [
    "tensogram>=0.18.1,<0.19",
    "tensogram-xarray>=0.18.1,<0.19",
    "earthkit-data>=0.19",
    "earthkit-utils",
    "numpy",
    "xarray>=2022.06",
]

[project.urls]
Homepage = "https://sites.ecmwf.int/docs/tensogram/main"
Repository = "https://github.com/ecmwf/tensogram"
Documentation = "https://sites.ecmwf.int/docs/tensogram/main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.entry-points."earthkit.data.sources"]
tensogram = "tensogram_earthkit.source"

[project.entry-points."earthkit.data.encoders"]
tensogram = "tensogram_earthkit.encoder"

[tool.hatch.build.targets.wheel]
packages = ["src/tensogram_earthkit"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "PT",    # flake8-pytest-style
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["RUF012"]

[tool.ruff.lint.isort]
known-third-party = ["numpy", "xarray", "tensogram", "earthkit", "pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
